Move Appearance settings to `menus_settings_appearance.cpp`
